WIP harvester endpoints update to schema

+++ prisma/migrations/20231005152443_initial/migration.sql
@@ -96,7 +96,7 @@ CREATE TABLE "HarvestOperation" (
     "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "start_time" TIMESTAMP(3),
     "end_time" TIMESTAMP(3),
-    "priorPeriodHarvested" INTEGER NOT NULL DEFAULT 0,
+    "priorPeriodHarvested" DOUBLE PRECISION NOT NULL DEFAULT 0.0,
 
     CONSTRAINT "HarvestOperation_pkey" PRIMARY KEY ("id")
 );
